US 11258 : added 'page not working link' to confirmation and cya screen on view claim (#204)
* added 'page not working link' to confirmation and cya screen on view claim

* pr comments addressed

* hook renamed and added to AccessibilityPage
