+++ src/components/override-sdk/infra/Assignment/Assignment.tsx
@@ -1,7 +1,6 @@
 import React, { useState, useEffect } from 'react';
 import PropTypes from 'prop-types';
 import { useTranslation } from 'react-i18next';
-import { getSdkConfig } from '@pega/react-sdk-components/lib/components/helpers/config_access';
 import { scrollToTop } from '../../../helpers/utils';
 import useAddErrorToPageTitle from '../../../helpers/hooks/useAddErrorToPageTitle';
 import ErrorSummary from '../../../BaseComponents/ErrorSummary/ErrorSummary';
@@ -49,19 +48,6 @@ export default function Assignment(props) {
   const isOnlyFieldDetails  = useIsOnlyField(null, children);// .isOnlyField;
   const [errorSummary, setErrorSummary] = useState(false);
   const [errorMessages, setErrorMessages] = useState<Array<OrderedErrorMessage>>([]);
-
-  const [referrerURL, setReferrerURL] = useState<string>(null);
-  const [hmrcURL, setHmrcURL] = useState<string>(null);
-
-  useEffect(() => {
-    const getReferrerURL = async () => {
-      const { serverConfig: { sdkContentServerUrl, sdkHmrcURL } } = await getSdkConfig();
-      setReferrerURL(sdkContentServerUrl);
-      setHmrcURL(sdkHmrcURL);
-    }
-    getReferrerURL();
-    scrollToTop();
-  }, []);
 
   const _containerName =  getPConnect().getContainerName();
   const context = getPConnect().getContextName();
@@ -305,9 +291,6 @@ export default function Assignment(props) {
             >
               {t("ASK_HMRC_ONLINE")} {t("OPENS_IN_NEW_TAB")}
             </a><br/><br/>
-            <a className="govuk-link" href={`${hmrcURL}contact/report-technical-problem?newTab=true&service=463&referrerUrl=${referrerURL}`} rel="noreferrer" target="_blank">
-                {t("PAGE_NOT_WORKING_PROPERLY")} {t("OPENS_IN_NEW_TAB")}
-            </a>
             </div>
           </div>
         </main>
