+++ src/samples/ChildBenefitsClaim/UserPortal.tsx
@@ -1,25 +1,13 @@
-import React,{ useEffect, useState } from 'react';
+import React from 'react';
 import Button from '../../components/BaseComponents/Button/Button';
 import { useTranslation } from 'react-i18next';
-import { getSdkConfig } from '@pega/react-sdk-components/lib/components/helpers/config_access';
+import useHMRCExternalLinks from '../../components/helpers/hooks/HMRCExternalLinks';
 
 
 export default function UserPortal(props) {
   const { beginClaim, children } = props;
   const { t } = useTranslation();
-
-  const [referrerURL, setReferrerURL] = useState<string>(null);
-  const [hmrcURL, setHmrcURL] = useState<string>(null);
-
-  useEffect(() => {
-    const getReferrerURL = async () => {
-      const { serverConfig: { sdkContentServerUrl, sdkHmrcURL } } = await getSdkConfig();
-      setReferrerURL(sdkContentServerUrl);
-      setHmrcURL(sdkHmrcURL);
-    }
-    getReferrerURL();
-    // scrollToTop();
-  }, []);
+  const {referrerURL, hmrcURL} = useHMRCExternalLinks();
 
   return (
     <>
@@ -69,8 +57,14 @@ export default function UserPortal(props) {
                   <li><span className='govuk-body'>{t('ANY_CHANGE_OF_CIRCUMSTANCE')}</span></li>
                 </ul>
                 <br/>
-                <a className="govuk-link" href={`${hmrcURL}contact/report-technical-problem?newTab=true&service=463&referrerUrl=${referrerURL}`} rel="noreferrer" target="_blank">
-                  {t("PAGE_NOT_WORKING_PROPERLY")} {t("OPENS_IN_NEW_TAB")}
+                <a
+                  lang='en'
+                  className='govuk-link hmrc-report-technical-issue '
+                  rel='noreferrer noopener'
+                  target='_blank'
+                  href={`${hmrcURL}contact/report-technical-problem?newTab=true&service=463&referrerUrl=${referrerURL}`}
+                >
+                  {t('PAGE_NOT_WORKING_PROPERLY')} {t("OPENS_IN_NEW_TAB")}
                 </a>
               </>
             </div>
