+++ src/samples/ChildBenefitsClaim/index.tsx
@@ -16,6 +16,7 @@ import AppHeader from '../../components/AppComponents/AppHeader';
 import AppFooter from '../../components/AppComponents/AppFooter';
 import LanguageToggle from '../../components/AppComponents/LanguageToggle';
 import LogoutPopup from '../../components/AppComponents/LogoutPopup';
+import useHMRCExternalLinks from '../../components/helpers/hooks/HMRCExternalLinks';
 
 import StartPage from './StartPage';
 import ConfirmationPage from './ConfirmationPage';
@@ -71,11 +72,12 @@ export default function ChildBenefitsClaim() {
   const [loadinginProgressClaims, setLoadingInProgressClaims] = useState(true);
   const [showSignoutModal, setShowSignoutModal] = useState(false);
   const [showTimeoutModal, setShowTimeoutModal] = useState(false)
-  const [authType, setAuthType] = useState('gg');  
+  const [authType, setAuthType] = useState('gg'); 
+  const {referrerURL, hmrcURL} = useHMRCExternalLinks();
   const history = useHistory();
 
   const { t } = useTranslation();
-  let operatorId = '';  
+  let operatorId = '';
   
   useEffect(()=> {
     setPageTitle();
@@ -489,6 +491,17 @@ export default function ChildBenefitsClaim() {
         <LanguageToggle PegaApp="true"/>
         <div id='pega-part-of-page'>
           <div id='pega-root'></div>
+          <div className='govuk-!-margin-bottom-8'>
+            <a
+              lang='en'
+              className='govuk-link hmrc-report-technical-issue '
+              rel='noreferrer noopener'
+              target='_blank'
+              href={`${hmrcURL}contact/report-technical-problem?newTab=true&service=463&referrerUrl=${referrerURL}`}
+            >
+              {t('PAGE_NOT_WORKING_PROPERLY')} {t("OPENS_IN_NEW_TAB")}
+            </a>
+          </div>
         </div>
 
         {showStartPage && <StartPage onStart={startNow} onBack={closeContainer} />}
